requirements.txt
aio-pika==9.3.1
jsondiff==2.0.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.23.2
python-dotenv==1.0.0
pydantic~=2.5.3
roslz4 --extra-index-url https://rospypi.github.io/simple/
